Move all the bundled dependencies to one location
thrift/lib/cl/externals/

The reasoning is this is going to be easier to maintain and we can
avoid duplication this way.

Bundled systems are loaded only when they're needed to build the
particular thing anyway, via ASDF. Fiasco won't be loaded when
building the tutorial binaries, for example - only its system
definition might, which is pretty harmless.